+++ packages/api/src/platforms/vtex/clients/fetch.ts
@@ -4,7 +4,7 @@ export const fetchAPI = async (info: RequestInfo, init?: RequestInit) => {
   const response = await fetch(info, init)
 
   if (response.ok) {
-    return response.json()
+    return response.status !== 204 ? response.json() : undefined
   }
 
   console.error(info, init, response)