+++ packages/sdk/src/analytics/events/signup.ts
@@ -1,12 +1,10 @@
 // This isn't an ecommerce exclusive event, but it makes sense to include it in stores
 
-import type { LocatorParam } from './common'
-
 export interface SignupParams {
   method?: string
 }
 
 export interface SignupEvent {
   name: 'signup'
-  params: SignupParams & LocatorParam
+  params: SignupParams
 }