+++ packages/sdk/src/analytics/events/view_item_list.ts
@@ -1,4 +1,4 @@
-import type { Item, LocatorParam } from './common'
+import type { Item } from './common'
 
 export interface ViewItemListParams<T extends Item = Item> {
   item_list_id?: string
@@ -8,5 +8,5 @@ export interface ViewItemListParams<T extends Item = Item> {
 
 export interface ViewItemListEvent<T extends Item = Item> {
   name: 'view_item_list'
-  params: ViewItemListParams<T> & LocatorParam
+  params: ViewItemListParams<T>
 }